Finish exercises 4.5-4.23

+++ part4/blog-list/controller/blog.js
@@ -1,30 +1,70 @@
 const blogRouter = require('express').Router()
 const Blog = require('../model/blog.js')
+const User = require('../model/user')
+const { tokenExtractor, userExtractor } = require('../utils/middleware')
 
-blogRouter.get('/', (request, response, next) => {
-  Blog.find({}).then((blogs) => {
-    response.json(blogs)
-  }).catch(e => next(e))
+blogRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate('user')
+  response.json(blogs)
 })
 
-blogRouter.post('/', (request, response, next) => {
-  const blog = new Blog(request.body)
+blogRouter.post('/', tokenExtractor, userExtractor,
+  async (request, response) => {
+    const user = request.user
 
-  blog.save().then((result) => {
-    response.status(201).json(result)
-  }).catch(e => next(e))
-})
+    const blog = new Blog({
+      title: request.body.title,
+      author: request.body.author || '',
+      url: request.body.url,
+      likes: request.body.likes,
+      user: user.userId,
+    })
+    const createdBlog = await blog.save()
 
-blogRouter.delete('/:id', async (req, res) => {
-  await Blog.deleteOne({ _id: req.params.id })
-  res.status(204).end()
-})
+    const userFromDB = await User.findById(user.userId)
 
-blogRouter.put('/:id', async (req, res) => {
+    console.log(userFromDB)
+    userFromDB.blogs.push(createdBlog._id)
+
+    await userFromDB.save()
+
+    response.status(201).json(createdBlog)
+
+  })
+
+blogRouter.delete('/:id', tokenExtractor, userExtractor,
+  async (req, res, next) => {
+    const blog = await Blog.findById(req.params.id)
+    if (!blog) {
+      res.status(204).end()
+      return
+    }
+
+    if (blog.user.toString() !== req.user.userId) {
+      const e = new Error('You do not own this blog.')
+      e.status = 403
+      return next(e)
+    }
+
+    await Blog.deleteOne({ _id: req.params.id })
+
+    const userDb = await User.findById(req.user.userId)
+
+    userDb.blogs = userDb.blogs.filter(
+      b => b.toString() !== req.params.id,
+    )
+
+    await userDb.save()
+
+    res.status(204).end()
+  })
+
+blogRouter.put('/:id', async (req, res, next) => {
   const blog = await Blog.findById(req.params.id)
   if (!blog) {
-    res.status(404).end()
-    return
+    const e = new Error('Requested blog does not exist.')
+    e.status = 400
+    return next(e)
   }
 
   if (req.body.title) {
